fix warning

+++ ml/deep_learning/autoencoder.hpp
@@ -305,7 +305,8 @@ public:
                 layers_.push_back(es);
             }else{                
                 if(layers_.size() <= i){
-                    layers_.emplace_back(TmpInput.rows(), params_.hidden_size_[i]);
+                    layers_.emplace_back(static_cast<int>(TmpInput.rows()),
+                                         params_.hidden_size_[i]);
                 }
                 reduce_cost(uni_int, re, Batch, TmpInput, layers_[i]);                
                 generate_activation(layers_[i], TmpInput,
